requirements.txt
absl-py==0.4.0
appnope==0.1.0
astor==0.7.1
atomicwrites==1.2.0
attrs==18.1.0
backcall==0.1.0
bleach==2.1.4
cycler==0.10.0
decorator==4.3.0
docopt==0.6.2
entrypoints==0.2.3
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.1
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.2
ptyprocess==0.6.0
py==1.5.4
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.7.2
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.1
torchvision==0.2.1
tornado==5.1
traitlets==4.3.2
unityagents==0.4.0
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.0
